search: prevent sending request with empty query
During model creation a request without query set is still send. This
is due to overloaded methods not being called during construction. So
this means TraktSearchModel::fetchMore isn't called during construction
and we'll need another way to block the first request.

This commit solves that by passing in a parameter specifying to not
send the request.