images: work around Trakt only returning only null URLs
When no image is available Trakt sometimes returns the the image, but
all the URL types are null. This happens for example for the poster of
a shows season when no poster is available.

This change checks if at least a single URL of, full, medium or thumb is
set. And the image will then only be set when that's the case.